ShinyApp: Add libraries for emeScheme and progress indicator
The switching of the scheme was extremely slow due to the installation of the dependant packages. THerefore, I added them to the ui.R file as they are included in the app and the loading is much faster.

A progress indicator tells the user what is happening when switching schemes.